pass in the signal now so either SIGKILL or SIGTERM can use this dialog.
2001-08-03  Kevin Vandersloot <kfv101@psu.edu>

        * procdialogs.c (create_kill_dialog): pass in the signal now
        so either SIGKILL or SIGTERM can use this dialog. Change the
        button labels accordingly.

        * callbacks.c: use the create_kill_dailog properly adding
        the signal to the argument.

        * interface.c: left align the disk name and mount point

        * util.c: fix a memory leak

+++ src/callbacks.c
@@ -262,7 +262,7 @@ popup_menu_end_process (GtkMenuItem *menuitem, gpointer data)
 		return;
 	
 	if (procdata->config.show_kill_warning)
-		procdialog_create_kill_dialog (procdata);
+		procdialog_create_kill_dialog (procdata, SIGTERM);
 	else
 		kill_process (procdata, SIGTERM);	
 }
@@ -274,8 +274,12 @@ popup_menu_kill_process (GtkMenuItem *menuitem, gpointer data)
 
         if (!procdata->selected_node)
 		return;
-	
-	kill_process (procdata, SIGKILL);	
+		
+	if (procdata->config.show_kill_warning)
+		procdialog_create_kill_dialog (procdata, SIGKILL);
+	else	
+		kill_process (procdata, SIGKILL);
+			
 }
 
 void 
@@ -324,7 +328,7 @@ cb_end_process_button_pressed          (GtkButton       *button,
 		return;
 		
 	if (procdata->config.show_kill_warning)
-		procdialog_create_kill_dialog (procdata);
+		procdialog_create_kill_dialog (procdata, SIGTERM);
 	else
 		kill_process (procdata, SIGTERM);
